requirements.txt
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
billiard==4.2.1
boto3==1.35.44
botocore==1.35.44
build==1.2.2.post1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
dill==0.3.6
dnslib==0.9.25
dnspython==2.7.0
ecdsa==0.19.0
eventlet==0.37.0
exceptiongroup==1.2.2
fastapi==0.115.2
flower==2.0.1
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
humanize==4.11.0
idna==3.10
importlib_metadata==8.5.0
jmespath==1.0.1
kombu==5.4.2
localstack==3.8.1
localstack-core==3.8.1
localstack-ext==3.8.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.1
plux==1.11.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyotp==2.9.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rich==13.9.2
rsa==4.9
s3transfer==0.10.3
semver==3.0.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
tabulate==0.9.0
tailer==0.4.1
tomli==2.0.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
zipp==3.20.2
